Fixed networking synchronisation

+++ src/main/java/com/redblueflame/herbocraft/ui/ReproducerBlockInterface.java
@@ -36,13 +36,13 @@ public class ReproducerBlockInterface extends BaseContainerScreen<ReproducerBloc
         // Progession Arrow Loading
         img = mainPanel.createChild(WHorizontalLoadingBar::new, Position.of(mainPanel, ((mainPanel.getWidth()) / 2) - 21, 55 - 7, 2), Size.of(14*2F, 14)).setParent(mainInterface);
         img.setTexture(new Identifier(HerboCraft.name, "textures/ui/arrow_filled.png"));
-        img.setState(container.entity.state);
+        img.setState((short) container.entity.state);
         mainPanel.add(base);
     }
 
     @Override
     public void render(MatrixStack matrices, int mouseX, int mouseY, float tickDelta) {
-        img.setState(getContainer().entity.state);
+        img.setState((short) getContainer().entity.state);
         super.render(matrices, mouseX, mouseY, tickDelta);
     }
